cmd: fix config and flags loading process
Get value from viper, like `viper.GetString()`

Value priority:

1. Return value from config file
2. Return value from flag value, passed by command line
3. Return value from flag default value
